STY: Add ECDF to distributions.empirical_distribution and rltd cleanup
