+++ CHANGES.txt
@@ -1,8 +1,8 @@
 trunk for 0.4.0
 ---------------
-* tools.tools.ECDF -> nonparametric.ECDF
-* tools.tools.monotone_fn_inverter -> nonparametric.monotone_fn_inverter
-* tools.tools.StepFunction -> nonparametric.StepFunction
+* tools.tools.ECDF -> distributions.ECDF
+* tools.tools.monotone_fn_inverter -> distributions.monotone_fn_inverter
+* tools.tools.StepFunction -> distributions.StepFunction
 
 trunk for 0.3.0
 ---------------
