+++ scikits/statsmodels/api.py
@@ -9,6 +9,7 @@ from .robust.robust_linear_model import RLM
 from .discrete.discrete_model import Poisson, Logit, Probit, MNLogit
 from .tsa import api as tsa
 import nonparametric
+import distributions
 from __init__ import test
 import version
 from info import __doc__
