+++ scikits/statsmodels/tools/tools.py
@@ -7,8 +7,8 @@ import numpy.lib.recfunctions as nprf
 import numpy.linalg as L
 from scipy.interpolate import interp1d
 from scipy.linalg import svdvals
-from scikits.statsmodels.nonparametric.distributions import (ecdf, 
-        monotone_fn_inverter, StepFunction)
+from scikits.statsmodels.distributions import (ECDF, monotone_fn_inverter, 
+                                               StepFunction)
 
 def _make_dictnames(tmp_arr, offset=0):
     """
@@ -352,14 +352,14 @@ def fullrank(X, r=None):
     return np.asarray(np.transpose(value)).astype(np.float64)
 
 StepFunction = np.deprecate(StepFunction, 
-                    old_name = 'scikits.statsmodels.tools.tools.StepFunction',
-                    new_name = 'scikits.statsmodels.nonparametric.StepFunction')
+                old_name = 'scikits.statsmodels.tools.tools.StepFunction',
+                new_name = 'scikits.statsmodels.distributions.StepFunction')
 monotone_fn_inverter = np.deprecate(monotone_fn_inverter, 
-                    old_name = 'scikits.statsmodels.tools.tools.monotone_fn_inverter',
-                    new_name = 'scikits.statsmodels.nonparametric.monotone_fn_inverter')
-ECDF = np.deprecate(ecdf, 
-                    old_name = 'scikits.statsmodels.tools.tools.ECDF',
-                    new_name = 'scikits.statsmodels.nonparametric.ECDF')
+                old_name = 'scikits.statsmodels.tools.tools.monotone_fn_inverter',
+                new_name = 'scikits.statsmodels.distributions.monotone_fn_inverter')
+ECDF = np.deprecate(ECDF, 
+                old_name = 'scikits.statsmodels.tools.tools.ECDF',
+                new_name = 'scikits.statsmodels.distributions.ECDF')
 
 
 def unsqueeze(data, axis, oldshape):
